# docs/swagger.yaml
info:
  contact: {}
paths:
...
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Realiza login com email e senha
      parameters:
      - description: E-mail
        in: formData
        name: email
        required: true
        type: string
      - description: Senha
        in: formData
        name: password
        required: true
        type: string
      produces:
      - text/html
      responses:
        "302":
          description: Redireciona para verificação de token
          schema:
            type: string
        "401":
          description: Credenciais inválidas
          schema:
            type: string
      summary: Login
      tags:
      - Autenticação
...
  /verificar-token:
    get:
      consumes:
      - text/html
      description: Exibe página para verificar o código enviado
      parameters:
      - description: E-mail
        in: query
        name: email
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: Exibe formulário de verificação
          schema:
            type: string
      summary: Página de verificação
      tags:
      - Autenticação
...
swagger: "2.0"